Add exception to isolate

+++ transpiler/transpiler/include/n1ql.h
@@ -56,17 +56,17 @@ struct InsertedCharsInfo {
   
   insert_type type;
   int type_len;
-  int64_t line_no;
-  int64_t index;
+  int32_t line_no;
+  int32_t index;
 };
 
 // Represents position of each char in the source code
 struct Pos {
   Pos() : line_no(0), col_no(0), index(0) {}
   
-  int64_t line_no;
-  int64_t col_no;
-  int64_t index;
+  int32_t line_no;
+  int32_t col_no;
+  int32_t index;
 };
 
 // Represents compilation status
@@ -75,9 +75,9 @@ struct CompilationInfo {
   
   std::string language;
   bool compile_success;
-  int64_t index;
-  int64_t line_no;
-  int64_t col_no;
+  int32_t index;
+  int32_t line_no;
+  int32_t col_no;
   std::string description;
 };
